Move test to simplify
